+++ src/emoji.cpp
@@ -1903,7 +1903,7 @@ EmojiEntityMapper::EmojiEntityMapper()
 {
   for (size_t i = 0; i < g_numEmojiEntities; i++)
   {
-    m_name2symGh.insert(std::make_pair(g_emojiEntities[i].name, i));
+    m_name2symGh.insert(std::make_pair(g_emojiEntities[i].name, static_cast<int>(i)));
   }
   for (size_t i = 0; i < g_numEmojiCompatibilityEntities; i++)
   {
